Add playGame() and relocate choice variables
Add playGame() function and repeat several times. This adds several
rounds of gameplay. Relocate humanChoice and computerChoice variables
inside playRound function so that they change each round to allow for
new round winners.

The game is able to repeat; however, the prompt box displays around
double the amount it is supposed. Every other score is logged in the
console. 5 scores log, but 10 rounds are played. This is the next bug
to be targeted.